+++ src/Plugin/PluginManagerInterface.php
@@ -15,4 +15,26 @@ interface PluginManagerInterface
      * @return \Para\Plugin\PluginInterface[]
      */
     public function fetchPluginsAvailable(): array;
+
+    /**
+     * Installs a plugin.
+     *
+     * @param string $name The name of the plugin.
+     * @param string $version The version of the plugin.
+     *
+     * @return string The output of the process that installs the plugin.
+     *
+     * @throws \Para\Exception\PluginAlreadyInstalledException When the plugin is already installed.
+     * @throws \Para\Exception\PluginNotFoundException When the plugin to install could not be found.
+     */
+    public function installPlugin(string $name, string $version): string;
+
+    /**
+     * Returns if a plugin is already installed.
+     *
+     * @param string $pluginName The name of the plugin.
+     *
+     * @return bool True if the plugin is already installed, otherwise false.
+     */
+    public function isInstalled(string $pluginName): bool;
 }