+++ tests/Unit/Plugin/PluginManagerTest.php
@@ -4,15 +4,21 @@ namespace Para\Tests\Unit\Plugin;
 
 use Composer\Composer;
 use Composer\Factory;
+use Composer\IO\NullIO;
+use Composer\Package\Locker;
 use Composer\Repository\ComposerRepository;
 use Composer\Repository\CompositeRepository;
 use Composer\Repository\RepositoryManager;
 use Para\Factory\CompositeRepositoryFactoryInterface;
 use Para\Factory\PluginFactoryInterface;
+use Para\Factory\ProcessFactoryInterface;
+use Para\Para;
 use Para\Plugin\PluginInterface;
 use Para\Plugin\PluginManager;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\Process\Process;
 
 /**
  * Class PluginManagerTest
@@ -50,6 +56,13 @@ class PluginManagerTest extends TestCase
     private $pluginFactory;
 
     /**
+     * The process factory mock object.
+     *
+     * @var \Para\Factory\ProcessFactoryInterface
+     */
+    private $processFactory;
+
+    /**
      * {@inheritdoc}
      */
     protected function setUp()
@@ -57,11 +70,13 @@ class PluginManagerTest extends TestCase
         $this->repositoryFactory = $this->prophesize(CompositeRepositoryFactoryInterface::class);
         $this->composerFactory = $this->prophesize(Factory::class);
         $this->pluginFactory = $this->prophesize(PluginFactoryInterface::class);
+        $this->processFactory = $this->prophesize(ProcessFactoryInterface::class);
 
         $this->pluginManager = new PluginManager(
             $this->repositoryFactory->reveal(),
             $this->composerFactory->reveal(),
             $this->pluginFactory->reveal(),
+            $this->processFactory->reveal(),
             'the/path/to/the/root/directory/of/para'
         );
     }
@@ -123,5 +138,128 @@ class PluginManagerTest extends TestCase
         $this->assertTrue(is_array($plugins), $plugins);
         $this->assertTrue($plugins['para-alias'] instanceof PluginInterface);
         $this->assertTrue($plugins['para-sync'] instanceof PluginInterface);
+    }
+
+    /**
+     * Tests that a plugin installs a plugin successfully
+     */
+    public function testInstallsAPluginSuccessfully()
+    {
+        $pluginName = 'lrackwitz/para-alias';
+        $pluginVersion = 'dev-master';
+
+        $commandline = 'composer require lrackwitz/para-alias dev-master';
+        $cwd = 'the/path/to/the/root/directory/of/para';
+
+        $locker = $this->prophesize(Locker::class);
+        $locker->getLockData()->shouldBeCalled();
+        $locker->getLockData()->willReturn([
+            'packages' => [],
+        ]);
+
+        $composer = $this->prophesize(Composer::class);
+        $composer->getLocker()->shouldBeCalled();
+        $composer->getLocker()->willReturn($locker->reveal());
+
+        $this->composerFactory
+            ->createComposer(Argument::any(), Argument::type('string'), false, Argument::type('string'), true)
+            ->shouldBeCalled();
+        $this->composerFactory
+            ->createComposer(Argument::any(), Argument::type('string'), false, Argument::type('string'), true)
+            ->willReturn($composer->reveal());
+
+        $process = $this->prophesize(Process::class);
+        $process->run()->shouldBeCalled();
+        $process->isSuccessful()->shouldBeCalled();
+        $process->isSuccessful()->willReturn(true);
+        $process->getOutput()->shouldBeCalled();
+        $process->getOutput()->willReturn('something');
+
+        $this->processFactory
+            ->getProcess($commandline, $cwd)
+            ->shouldBeCalled();
+        $this->processFactory
+            ->getProcess($commandline, $cwd)
+            ->willReturn($process->reveal());
+
+        $this->pluginManager->installPlugin($pluginName, $pluginVersion);
+    }
+
+    /**
+     * Tests that installing a plugin fails.
+     *
+     * @expectedException \Para\Exception\PluginNotFoundException
+     */
+    public function testInstallingAPluginFails()
+    {
+        $pluginName = 'lrackwitz/para-alias';
+        $pluginVersion = 'dev-master';
+
+        $commandline = 'composer require lrackwitz/para-alias dev-master';
+        $cwd = 'the/path/to/the/root/directory/of/para';
+
+        $locker = $this->prophesize(Locker::class);
+        $locker->getLockData()->shouldBeCalled();
+        $locker->getLockData()->willReturn([
+            'packages' => [],
+        ]);
+
+        $composer = $this->prophesize(Composer::class);
+        $composer->getLocker()->shouldBeCalled();
+        $composer->getLocker()->willReturn($locker->reveal());
+
+        $this->composerFactory
+            ->createComposer(Argument::any(), Argument::type('string'), false, Argument::type('string'), true)
+            ->shouldBeCalled();
+        $this->composerFactory
+            ->createComposer(Argument::any(), Argument::type('string'), false, Argument::type('string'), true)
+            ->willReturn($composer->reveal());
+
+        $process = $this->prophesize(Process::class);
+        $process->run()->shouldBeCalled();
+        $process->isSuccessful()->shouldBeCalled();
+        $process->isSuccessful()->willReturn(false);
+
+        $this->processFactory
+            ->getProcess($commandline, $cwd)
+            ->shouldBeCalled();
+        $this->processFactory
+            ->getProcess($commandline, $cwd)
+            ->willReturn($process->reveal());
+
+        $this->pluginManager->installPlugin($pluginName, $pluginVersion);
+    }
+
+    /**
+     * Tests that the isInstalled() method returns true when the plugin is already installed.
+     */
+    public function testTheIsInstalledMethodReturnsTrueWhenThePluginIsAlreadyInstalled()
+    {
+        $pluginName = 'lrackwitz/para-alias';
+
+        $locker = $this->prophesize(Locker::class);
+        $locker->getLockData()->shouldBeCalled();
+        $locker->getLockData()->willReturn([
+            'packages' => [
+                [
+                    'name' => 'lrackwitz/para-alias',
+                ],
+            ],
+        ]);
+
+        $composer = $this->prophesize(Composer::class);
+        $composer->getLocker()->shouldBeCalled();
+        $composer->getLocker()->willReturn($locker->reveal());
+
+        $this->composerFactory
+            ->createComposer(Argument::any(), Argument::type('string'), false, Argument::type('string'), true)
+            ->shouldBeCalled();
+        $this->composerFactory
+            ->createComposer(Argument::any(), Argument::type('string'), false, Argument::type('string'), true)
+            ->willReturn($composer->reveal());
+
+        $result = $this->pluginManager->isInstalled($pluginName);
+
+        $this->assertTrue($result);
     }
 }